refactor(web): port AdminDateRangeSelector to refresh-components
