+++ web/src/app/ee/admin/performance/query-history/QueryHistoryTable.tsx
@@ -17,7 +17,7 @@ import { Feedback, TaskStatus } from "@/lib/types";
 import {
   DateRange,
   AdminDateRangeSelector,
-} from "@/components/dateRangeSelectors/AdminDateRangeSelector";
+} from "@/refresh-components/AdminDateRangeSelector";
 import { PageSelector } from "@/components/PageSelector";
 import Link from "next/link";
 import type { Route } from "next";
