+++ web/src/app/ee/admin/performance/query-history/utils.ts
@@ -1,4 +1,4 @@
-import { DateRange } from "../../../../../components/dateRangeSelectors/AdminDateRangeSelector";
+import { DateRange } from "@/refresh-components/AdminDateRangeSelector";
 import { START_QUERY_HISTORY_EXPORT_URL } from "./constants";
 
 export const withRequestId = (url: string, requestId: string): string =>
