+++ web/src/views/admin/WorkspaceAnalyticsPage/AnalyticsChart.tsx
@@ -5,7 +5,7 @@ import { PageLoader, Section } from "@opal/layouts";
 import type { RichStr } from "@opal/types";
 import AreaChart from "@/refresh-components/AreaChart";
 import { getDatesList } from "@/app/ee/admin/performance/lib";
-import { DateRangePickerValue } from "@/components/dateRangeSelectors/AdminDateRangeSelector";
+import { DateRangePickerValue } from "@/refresh-components/AdminDateRangeSelector";
 import {
   ChartSeries,
   ChartState,
