+++ web/src/views/admin/WorkspaceAnalyticsPage/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { AdminDateRangeSelector } from "@/components/dateRangeSelectors/AdminDateRangeSelector";
+import { AdminDateRangeSelector } from "@/refresh-components/AdminDateRangeSelector";
 import { useTimeRange } from "@/app/ee/admin/performance/lib";
 import {
   FeedbackChart,
